+++ pythainlp/__init__.py
@@ -55,6 +55,7 @@ thai_pangram: str = """กีฬาบังลังก์ ฿๑,๒๓๔,๕
 __all__: list[str] = [
     "collate",
     "correct",
+    "is_offline_mode",
     "pos_tag",
     "romanize",
     "spell",
@@ -76,5 +77,6 @@ from pythainlp.tokenize import (
     subword_tokenize,
     word_tokenize,
 )
+from pythainlp.tools.path import is_offline_mode
 from pythainlp.transliterate import romanize, transliterate
 from pythainlp.util import collate, thai_strftime